refactor(frontend): WorkflowActivityList -> SimpleActivityList
