+++ frontend/src/features/dataset/state/datasetActions.ts
@@ -53,6 +53,8 @@ function fetchBody (ref: QriRef, page: number, pageSize: number): ApiAction {
         pageSize
       },
       segments: {
+        // TODO(b5): this is a hack for now while we don't have session support,
+        // restore ASAP
         // peername: ref.username,
         peername: 'me',
         name: ref.name
