+++ frontend/src/features/deploy/state/deployState.ts
@@ -17,8 +17,13 @@ const initialState: DeployState = {
 
 export const deployReducer = createReducer(initialState, {
   // on successful workflow fetch, derive deploy state
+  'API_WORKFLOW_SUCCESS': (state: DeployState, action) => {
+    const w = action.payload.data as Workflow
+    state.status[w.id] = workflowDeployStatus(w)
+  },
+  // on successful deployment, derive deploy state
   'API_DEPLOY_SUCCESS': (state: DeployState, action) => {
     const w = action.payload.data.workflow as Workflow
     state.status[w.id] = workflowDeployStatus(w)
-  }
+  },
 })
