+++ frontend/src/features/session/modal/LogInModal.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react'
-import { Redirect } from 'react-router'
 import { Link } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 import { SyncLoader } from 'react-spinners'
