+++ frontend/src/features/session/modal/SignUpModal.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react'
-import { Redirect } from 'react-router'
 import { useDispatch, useSelector } from 'react-redux';
 import { SyncLoader } from 'react-spinners'
